% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_axes.R
\name{add_axes}
\alias{add_axes}
\title{Add axes to 3D rayshader plot}
\usage{
add_axes(ras, zscale, axis.col = 1, ...)
}
\arguments{
\item{ras}{raster object with bathymetry. This is the raster used to create the hillshade nad heightmap of the rayshader 3D plot
tag detection data, metadata and station information}

\item{zscale}{scaling factor for the depth, indicating depth exaggeration}

\item{axis.col}{color of axis and axis text, defaulted to 1}

\item{...}{additional arguments for \code{\link{bbox3d}} function}
}
\value{
Adds axes to an existing 3D rayshader plot
}
\description{
Helper function to add axes to a 3D rayshader plot
}
\examples{
TBD

}
