% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fkud.R
\name{add_fkud}
\alias{add_fkud}
\title{Add fixed Kernel Utilisation Distribution estimates to 3D rayshader plot}
\usage{
add_fkud(ras, det, zscale, lonlat = FALSE, mul = 1, ...)
}
\arguments{
\item{ras}{raster object with bathymetry. This is the raster used to create the hillshade nad heightmap of the rayshader 3D plot
tag detection data, metadata and station information}

\item{det}{...}

\item{zscale}{scaling factor for the depth, indicating depth exaggeration}

\item{lonlat}{...}

\item{mul}{...}

\item{...}{additional arguments for \code{\link{plot.kde}} function}
}
\value{
Adds 3DKUD voxels to an existing 3D rayshader plot
}
\description{
Helper function to add 3DKUD a 3D rayshader plot
}
\examples{
TBD

}
