% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_points.R
\name{add_points}
\alias{add_points}
\title{Add points or lines to a 3D rayshader plot}
\usage{
add_points(ras, det, zscale, lonlat = FALSE, col = "red", alpha = 0.8,
  line = F, ...)
}
\arguments{
\item{ras}{raster object with bathymetry. This is the raster used to create the hillshade nad heightmap of the rayshader 3D plot
tag detection data, metadata and station information}

\item{det}{...}

\item{zscale}{scaling factor for the depth, indicating depth exaggeration}

\item{lonlat}{...}

\item{col}{colour of point or line}

\item{alpha}{transparency of point or line}

\item{line}{logical if to plot lines instead of points}

\item{...}{additional arguments for \code{\link{points3d}} or \code{\link{lines3d}} function depending on what your plotting}
}
\value{
Adds points or lines to an existing 3D rayshader plot
}
\description{
Helper function to add points or lines a 3D rayshader plot
}
\examples{
TBD

}
