% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bath.R
\name{plot_bath}
\alias{plot_bath}
\title{Plot bathymetry with transparency control}
\usage{
plot_bath(hillshade, heightmap, zscale = 1, water = FALSE, waterdepth = 0,
  watercolor = "lightblue", wateralpha = 0.5, theta = 45, phi = 45,
  fov = 0, zoom = 1, background = "white", windowsize = c(600, 600),
  ...)
}
\arguments{
\item{hillshade}{Hillshade/image to be added to 3D surface map.}

\item{heightmap}{A two-dimensional matrix, where each entry in the matrix is the elevation at that point. All points are assumed to be evenly spaced.}

\item{zscale}{Default `1`. The ratio between the x and y spacing (which are assumed to be equal) and the z axis. For example, if the elevation levels are in units
of 1 meter and the grid values are separated by 10 meters, `zscale` would be 10. Adjust the zscale down to exaggerate elevation features.}

\item{water}{Default `FALSE`. If `TRUE`, a water layer is rendered.}

\item{waterdepth}{Default `0`. Water level.}

\item{watercolor}{Default `lightblue`. Color of the water.}

\item{wateralpha}{Default `0.5`. Water transparency.}

\item{theta}{Default `45`. Rotation around z-axis.}

\item{phi}{Default `45`. Azimuth angle.}

\item{fov}{Default `0`--isometric. Field-of-view angle.}

\item{zoom}{Default `1`. Zoom factor.}

\item{background}{Default `grey10`. Color of the background.}

\item{windowsize}{Default `c(600,600)`. Width and height of the `rgl` device displaying the plot.}

\item{...}{Additional arguments to pass to the `rgl::par3d` function.}

\item{baseshape}{Default `rectangle`. Shape of the base. Options are c("rectangle","circle","hex").}

\item{solid}{Default `TRUE`. If `FALSE`, just the surface is rendered.}

\item{soliddepth}{Default `auto`, which sets it to the lowest elevation in the matrix minus one unit (scaled by zscale). Depth of the solid base.}

\item{solidcolor}{Default `grey20`. Base color.}

\item{solidlinecolor}{Default `grey30`. Base edge line color.}

\item{shadow}{Default `TRUE`. If `FALSE`, no shadow is rendered.}

\item{shadowdepth}{Default `auto`, which sets it to `soliddepth - soliddepth/10`. Depth of the shadow layer.}

\item{shadowcolor}{Default `grey50`. Color of the shadow.}

\item{shadowwidth}{Default `auto`, which sizes it to 1/10th the smallest dimension of `heightmap`. Width of the shadow in units of the matrix.}

\item{waterlinecolor}{Default `NULL`. Color of the lines around the edges of the water layer.}

\item{waterlinealpha}{Default `1`. Water line tranparency.}

\item{linewidth}{Default `2`. Width of the edge lines in the scene.}

\item{lineantialias}{Default `FALSE`. Whether to anti-alias the lines in the scene.}
}
\description{
Displays the shaded map in 3D with the `rgl` package. Altered version of the \code{\link{plot_3d}} function to allow adjustment of transparency
}
\examples{
TBD

}
