% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol3d.R
\name{vol3d}
\alias{vol3d}
\title{Calculate 3DKUD volume}
\usage{
vol3d(fhat, cont = NULL)
}
\arguments{
\item{fhat}{a kernel density estimate object of class `kde`. see \code{\link{?kde}} using the `ks` package}

\item{cont}{voxel contour for which volume will be calulated (e.g. 50, 95)}
}
\value{
A numeric vector with volume of kde object
}
\description{
Helper function to calculate 3DKUD volume for a `kde` object
}
\examples{
## Import detection data
data(exampletag)

## Estimate 3D kernel density using the ks package
library(ks)
H.pi <- Hpi(exampletag, binned = TRUE) * 3
fhat <- kde(exampletag, H = H.pi)

## Calculate 3D KUD volume
vol3d(fhat = fhat, cont = 50)   ## volume of 50\% contour in m3
vol3d(fhat = fhat, cont = 95)   ## volume of 95\% contour in m3


}
